+++ lib/view/main_page_layout.dart
@@ -5,24 +5,10 @@ import 'package:flutter_svg/svg.dart';
 import 'package:flutter_training/components/dialog/error_message_dialog.dart';
 import 'package:flutter_training/constant/weather_condition.dart';
 import 'package:flutter_training/gen/assets.gen.dart';
+import 'package:flutter_training/provider/weather_info_notifier.dart';
 import 'package:flutter_training/repository/fetch_yumemi_weather.dart';
 import 'package:flutter_training/repository/result.dart';
 import 'package:flutter_training/view_model/weather_info.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'main_page_layout.g.dart';
-
-@riverpod
-class WeatherInfoNotifier extends _$WeatherInfoNotifier {
-  @override
-  WeatherInfo? build(WeatherInfo? weatherInfo) {
-    return weatherInfo;
-  }
-
-  Future<void> update(WeatherInfo newWeatherInfo) async {
-    state = newWeatherInfo;
-  }
-}
 
 /// 大枠のウィジェット
 class MainPageLayout extends StatelessWidget {
